requirements.txt
flask==3.0.3
flask-cors==4.0.1
numpy==1.26.4
opencv-python-headless==4.10.0.84
tensorflow-cpu==2.16.1
torch==2.3.1
torchvision==0.18.1
torchaudio==2.3.1
transformers==4.43.3
pillow==10.4.0
matplotlib==3.9.1
